requirements.txt
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10
mysql-connector-python==9.4.0
requests==2.32.4
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
